chore[d64]: Added event example
