--- HealthcareSystem/EventSourcingService/Service/IEventStoreService.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using EventSourcingService.Model;
-
-namespace EventSourcingService.Service
-{
-    public interface IEventStoreService
-    {
-        public IEnumerable<DomainEvent> GetAllEvents();
-
-        public void Add(DomainEvent domainEvent);
-    }
-}
